// src/services/orders-table/user-orders-table/types.ts
export enum UserOrdersDataActionTypes {
  USER_ORDER_TABLE_WS_MESSAGE = 'USER_ORDER_TABLE_WS_MESSAGE',
  USER_ORDER_TABLE_DISCONNECT = 'USER_ORDER_TABLE_DISCONNECT',
  USER_ORDER_TABLE_WS_CONNECTING = 'USER_ORDER_TABLE_WS_CONNECTING',
  USER_ORDER_TABLE_WS_OPEN = 'USER_ORDER_TABLE_WS_OPEN',
  USER_ORDER_TABLE_WS_CLOSE = 'USER_ORDER_TABLE_WS_CLOSE',
  USER_ORDER_TABLE_WS_ERROR = 'USER_ORDER_TABLE_WS_ERROR',
  USER_ORDER_TABLE_CONNECT = 'USER_ORDER_TABLE_CONNECT',
}